requirements.txt
asgiref==3.5.0
coverage==6.3.2
Django==4.0.3
django-crispy-forms==1.14.0
psycopg2==2.9.3
pytz==2020.1
sqlparse==0.3.1
tzdata==2022.1
djangorestframework==3.13.1
Markdown==3.3.6
django-filter==21.1
